requirements.txt
cmake 
dlib 
face_recognition
fastapi
uvicorn
python-multipart
pillow
numpy